+++ fs/lib/src/common/fs_extension.dart
@@ -1,4 +1,5 @@
 import 'package:fs_shim/fs.dart';
+import 'package:meta/meta.dart';
 
 import 'fs_sandbox.dart';
 
@@ -19,10 +20,27 @@ extension FsShimFileSystemExtension on FileSystem {
 
   /// Absolute path
   String absolutePath(String path) {
+    path = removeDotStart(path);
+    if (path == '.') {
+      return currentDirectory.path;
+    }
     if (this.path.isAbsolute(path)) {
       return path;
     } else {
       return this.path.join(currentDirectory.path, path);
     }
+  }
+}
+
+/// File system extension (private)
+@internal
+extension FsShimFileSystemPrvExtension on FileSystem {
+  String removeDotStart(String path) {
+    var sep = this.path.separator;
+    var dotStart = '.$sep';
+    while (path.startsWith(dotStart)) {
+      path = path.substring(dotStart.length);
+    }
+    return path;
   }
 }